fix: drift build settings and `make migration`
